do not move dev setup scripts yet
