+++ src/tasks/index.ts
@@ -12,3 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 export * from './tasks'
+export * from './setup'
