--- tests/devsetup/index.ts
@@ -1,2 +1,0 @@
-export * from './env'
-export * from './utils'
