+++ tests/tasks/tasks.test.ts
@@ -12,7 +12,7 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-import { getSendAmount, getStakeAmount, sendFunds } from '../devsetup'
+import { getSendAmount, getStakeAmount, sendFunds } from '../../src/tasks/setup'
 import { Tasks } from '../../src/tasks'
 
 import chai from 'chai'
